Refactor TransactionClient to use pydantic types
