+++ tests/test_paystack.py
@@ -48,8 +48,8 @@ from pypaystack2.sub_clients.async_clients.splits import AsyncTransactionSplitCl
 from pypaystack2.sub_clients.async_clients.subaccounts import AsyncSubAccountClient
 from pypaystack2.sub_clients.async_clients.subscriptions import AsyncSubscriptionClient
 from pypaystack2.sub_clients.async_clients.terminals import AsyncTerminalClient
+from pypaystack2.sub_clients.async_clients.transactions import AsyncTransactionClient
 from pypaystack2.sub_clients.sync_clients.products import AsyncProductClient
-from pypaystack2.sub_clients.transactions import AsyncTransactionClient
 from pypaystack2.sub_clients.transfer_recipients import (
     TransferRecipientClient,
     AsyncTransferRecipientClient,
